Meet up/meetup/requirements.txt
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2